Attempt to Make Goal Creations Work
In this change, I tried to make the input values for the "Create Goals"
actually append a new table for a new goal and subgoals. The problem I
ran into though is this: when I make jquery that states input.value =
new goaldiv.value, the divs for the "New Goals" have to have different
names each time a new goal needs to be made. I will have to consult Tom.